move preprocessing functions to common preprocessing util file
